requirements_test.txt
# ── Core dependencies ──────────────────────────────────────────────
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary>=2.9.7
twilio==9.0.4
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
Pillow==11.0.0

# ── Testing ────────────────────────────────────────────────────────
pytest>=8.3.0
pytest-django>=4.9.0
pytest-cov>=5.0.0
pytest-xdist>=3.6.0          # Parallel test execution
coverage[toml]>=7.6.0

# ── Code quality / CI ──────────────────────────────────────────────
flake8>=7.0.0
flake8-bugbear>=24.0.0
safety>=3.2.0
